+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Nested `z.array().nonempty()` failed to migrate. They are properly handled now. [#63](https://github.com/nicoespeon/zod-v3-to-v4/issues/63)
+
 ## [1.10.0] - 2025-12-15
 
 ### Added