+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.11.0] - 2025-12-30
+
 ### Added
 
 - The codemod now supports classic monorepo structures with a top-level `tsconfig.json` that [has references to others](https://www.typescriptlang.org/docs/handbook/project-references.html). [#62](https://github.com/nicoespeon/zod-v3-to-v4/issues/62)
@@ -272,7 +274,8 @@ For more information about Zod v4 changes, see the [official migration guide](ht
 
 <!-- Links -->
 
-[Unreleased]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.10.0...HEAD
+[Unreleased]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.11.0...HEAD
+[1.11.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.10.0...1.11.0
 [1.10.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.9.0...1.10.0
 [1.9.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.8.0...1.9.0
 [1.8.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.7.1...1.8.0